+++ App/Client/Services/ClocksService.cs
@@ -15,7 +15,6 @@ namespace Tellurian.Trains.MeetingApp.Client.Services
         }
 
         private readonly HttpClient Http;
-        private static string ApiKey => ClockSettings.ClockApiKey;
 
         public async Task<IEnumerable<string>> AvailableClocks() =>
             await Http.GetFromJsonAsync<IEnumerable<string>>("api/clocks/available").ConfigureAwait(false);
@@ -25,7 +24,7 @@ namespace Tellurian.Trains.MeetingApp.Client.Services
             if (string.IsNullOrEmpty(clockName)) return Array.Empty<ClockUser>();
             try
             {
-                return await Http.GetFromJsonAsync<IEnumerable<ClockUser>>($"api/clocks/{clockName}/users?apiKey={ApiKey}&password={administratorPassword}").ConfigureAwait(false);
+                return await Http.GetFromJsonAsync<IEnumerable<ClockUser>>($"api/clocks/{clockName}/users?password={administratorPassword}").ConfigureAwait(false);
             }
             catch (HttpRequestException)
             {
@@ -46,16 +45,16 @@ namespace Tellurian.Trains.MeetingApp.Client.Services
         }
 
         public async Task<HttpResponseMessage> Start(string clockName, string? clockPassword, string? userName) =>
-            await Http.PutAsync($"api/clocks/{clockName}/start?apiKey={ApiKey}&user={userName}&password={clockPassword}", null).ConfigureAwait(false);
+            await Http.PutAsync($"api/clocks/{clockName}/start?user={userName}&password={clockPassword}", null).ConfigureAwait(false);
 
         public async Task<HttpResponseMessage> Stop(string clockName, string? clockPassword, string? userName, string stopReason) =>
-            await Http.PutAsync($"api/clocks/{clockName}/stop?apiKey={ApiKey}&user={userName}&password={clockPassword}&reason={stopReason}", null).ConfigureAwait(false);
+            await Http.PutAsync($"api/clocks/{clockName}/stop?user={userName}&password={clockPassword}&reason={stopReason}", null).ConfigureAwait(false);
 
-        public async Task<HttpResponseMessage> User(string clockName, string? userName) =>
-            await Http.PutAsync($"api/clocks/{clockName}/user?apiKey={ApiKey}&user={userName}&client={ClockStatusExtension.ClientVersionNumber}", null).ConfigureAwait(false);
+        public async Task<HttpResponseMessage> User(string clockName, string? clockPassword, string? userName) =>
+            await Http.PutAsync($"api/clocks/{clockName}/user?user={userName}&password={clockPassword}&client={ClockStatusExtension.ClientVersionNumber}", null).ConfigureAwait(false);
 
         public async Task<HttpResponseMessage> Update(string clockName, string? userName, string? administratorPassword,  ClockSettings settings) =>
-            await Http.PostAsJsonAsync($"api/clocks/{clockName}/update?apiKey={ApiKey}&user={userName}&password={administratorPassword}", settings).ConfigureAwait(false);
+            await Http.PostAsJsonAsync($"api/clocks/{clockName}/update?user={userName}&password={administratorPassword}", settings).ConfigureAwait(false);
 
         public async Task<ClockSettings> GetSettings(string clockName) =>
             await Http.GetFromJsonAsync<ClockSettings>($"api/clocks/{clockName}/Settings").ConfigureAwait(false);